billy-b-assistant/requirements.txt
sounddevice
websockets
numpy
scipy
gpiozero
python-dotenv
pydub
paho-mqtt
requests
openai